Use environment defaults for initial connection
As far as I know, if you don't specify any options to the connection then libpq will pull defaults out of the environment, or as a last resort use it's built-in defaults. I think that's better than forcing localhost and 5432.

The localhost part is particularly annoying, since frequently IP connections require passwords but unix socket ones don't.

Note that I haven't tested this.